requirements.txt
# Backend dependencies
fastapi==0.104.1
uvicorn==0.23.2
python-multipart==0.0.6
pydantic==2.4.2
python-dotenv==1.0.0
jinja2==3.1.2

# OCR dependencies
# Note: Qari-OCR would need to be installed separately as it's a local library
# Install with: pip install git+https://github.com/mush42/qari-ocr.git

# PDF processing
pdf2image==1.16.3
# Note: poppler-utils is a system package, not a Python package
# Install manually: 
# - Windows: Download from https://github.com/oschwartz10612/poppler-windows/releases/
# - macOS: brew install poppler
# - Linux: apt-get install poppler-utils

# Document export
python-docx==1.0.1

# Translation and transliteration
openai==1.3.5
# Using a different translation library to avoid dependency conflicts
translators==5.8.2
arabic-reshaper==3.0.0
python-bidi==0.4.2

# For development
pylint==3.0.2
black==23.11.0